refactor: decompose plugins
